Introduce helper file to services
