+++ frontend/src/services/applicationService.js
@@ -2,54 +2,7 @@
  * Application Service
  * Handles all API interactions related to application management
  */
-import { checkTokenExpiration } from "./auth.service.js";
-
-/**
- * Base API request handler with common configuration
- * @param {string} endpoint - API endpoint
- * @param {string} method - HTTP method
- * @param {Object} [body] - Request body (optional)
- * @returns {Promise<Object>} - API response data
- * @throws {Error} - If request fails
- */
-async function makeApiRequest(endpoint, method, body = null) {
-  const headers = {
-    "Content-Type": "application/json",
-    'Authorization': `Bearer ${sessionStorage.getItem('token')}`,
-  };
-
-  const requestOptions = {
-    method,
-    headers,
-  };
-
-  if (body) {
-    requestOptions.body = JSON.stringify(body);
-  }
-
-  const response = await fetch(endpoint, requestOptions);
-  checkTokenExpiration(response);
-
-  const responseJson = await response.json();
-  if (!responseJson.success) {
-    throw new Error(responseJson.message || `Failed to ${method.toLowerCase()} ${endpoint}`);
-  }
-
-  return responseJson.data;
-}
-
-/**
- * Handles API request errors consistently
- * @param {Error} error - The caught error 
- * @param {string} errorMessage - Custom error message
- * @returns {Object} - Standardized error response
- */
-function handleApiError(error, errorMessage) {
-  if (error.response && error.response.status === 403) {
-    return { status: 403, message: "You are not authorized to perform this action" };
-  }
-  return { status: 500, message: errorMessage };
-}
+import { makeApiRequest, handleApiError } from './helper';
 
 /**
  * Retrieves all applications for the current user
